requirements.txt
huawei_solar
pysunspec2
serial
pymodbus